+++ ultralytics/nn/modules/head.py
@@ -1875,68 +1875,3 @@ class SemanticSegment(nn.Module):
                 return cls.to(torch.uint8 if self.nc <= 256 else torch.int32)
             return y
         return logits
-
-class AnchorFreeHead(Detect):
-    def __init__(self, ch, nc):
-        super().__init__(nc, ch=(ch,))
-
-        self.nc = nc
-
-        self.bbox = nn.Conv2d(
-            ch,
-            4,
-            kernel_size=1
-        )
-
-        self.cls = nn.Conv2d(
-            ch,
-            nc,
-            kernel_size=1
-        )
-
-        self.center = nn.Conv2d(
-            ch,
-            1,
-            kernel_size=1
-        )
-    def forward(self, x):
-        bbox = self.bbox(x)
-        cls = self.cls(x)
-        center = self.center(x)
-
-        return torch.cat(
-            [
-                bbox,
-                cls,
-                center
-            ],
-            dim=1
-        )
-
-class HybridHead(nn.Module):
-    def __init__(self, nc, ch):
-        super().__init__()
-
-        self.anchor_head = Detect(nc, ch)
-
-        self.af_head = nn.ModuleList(
-            [
-                AnchorFreeHead(
-                    c,
-                    nc
-                )
-                for c in ch
-            ]
-        )
-    
-    def forward(self, x):
-        anchor_output = self.anchor_head(x)
-        af_output = []
-        for i, f in enumerate(x):
-            af_output.append(
-                self.af_head[i](f)
-            )
-        return {
-            "anchor": anchor_output,
-            "anchor_free": af_output
-        }
